bump version to 3.05, and fix for Issue 42
Includes a fix for Issue 42 except that Slate mtl editor shows a zero for the map channel initially (but is correct in the dialog box).

Also a fix for XYZ_Tiling, and the "random" checkboxes. Parts of the code annotated for future maintainers.

Summary of changes in change notes.txt